Split the logger into two interfaces
Having the `start` and `stop` member functions on the public `logger`
interface semantically makes no sense and they are only there because
the `actor_system` needs them. We "hide" those methods by making them
private and then tag the `actor_system` as a friend. However, just
having an internal interface that adds those methods that the actor
system then uses separates concerns much better.

Furthermore, the `init` method could be removed entirely by simply
initializing all required state in the constructor directly.
