+++ libcaf_core/CMakeLists.txt
@@ -152,6 +152,7 @@ caf_add_component(
     caf/detail/abstract_worker_hub.cpp
     caf/detail/actor_system_access.cpp
     caf/detail/actor_system_config_access.cpp
+    caf/detail/asynchronous_logger.cpp
     caf/detail/aligned_alloc.cpp
     caf/detail/assert.cpp
     caf/detail/atomic_ref_counted.cpp
