+++ libcaf_test/caf/test/reporter.hpp
@@ -6,6 +6,7 @@
 
 #include "caf/test/fwd.hpp"
 
+#include "caf/detail/asynchronous_logger.hpp"
 #include "caf/detail/log_level.hpp"
 #include "caf/detail/test_export.hpp"
 #include "caf/format_string_with_location.hpp"
@@ -122,7 +123,7 @@ public:
   static std::unique_ptr<reporter> make_default();
 
   /// Creates a logger that forwards events to the current reporter.
-  static intrusive_ptr<logger> make_logger();
+  static intrusive_ptr<detail::asynchronous_logger> make_logger();
 };
 
 } // namespace caf::test
